Makes code changes necessary to support version 1.2.0.4274 of MongoDB Official Driver

+++ src/Elmah.MongoDB/MongoErrorLog.cs
@@ -211,7 +211,7 @@ namespace Elmah
 				errorEntryList.Add(new ErrorLogEntry(this, id, error));
 			}
 
-			return _collection.Count();
+			return (int) _collection.Count();
 		}
 
 		public static int GetCollectionLimit(IDictionary config)
